# docs/swagger.yaml
basePath: /api
definitions:
...
  rip_internal_app_dto.AuthResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      message:
        type: string
      refresh_token:
        type: string
      token_type:
        type: string
      user:
        $ref: '#/definitions/rip_internal_app_dto.UserResponse'
    type: object
  rip_internal_app_dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  rip_internal_app_dto.MessageResponse:
    properties:
      message:
        type: string
    type: object
  rip_internal_app_dto.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
...
  rip_internal_app_dto.SignInRequest:
    properties:
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
  rip_internal_app_dto.SignUpRequest:
    properties:
      login:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - login
    - password
    type: object
...
  rip_internal_app_dto.UpdateUserRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  rip_internal_app_dto.UserResponse:
    properties:
      is_moderator:
        type: boolean
      login:
        type: string
      user_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Route Information Platform API
  title: RIP API
  version: "1.0"
paths:
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get new access token using refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rip_internal_app_dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rip_internal_app_dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rip_internal_app_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rip_internal_app_dto.ErrorResponse'
      summary: Refresh JWT tokens
      tags:
      - auth
  /auth/signin:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT tokens
      parameters:
      - description: User credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rip_internal_app_dto.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rip_internal_app_dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rip_internal_app_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rip_internal_app_dto.ErrorResponse'
      summary: User authentication
      tags:
      - auth
  /auth/signout:
    post:
      consumes:
      - application/json
      description: Invalidate user tokens
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rip_internal_app_dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rip_internal_app_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rip_internal_app_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: User logout
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rip_internal_app_dto.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/rip_internal_app_dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rip_internal_app_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rip_internal_app_dto.ErrorResponse'
      summary: User registration
      tags:
      - auth
...
  /users/me:
    get:
      consumes:
      - application/json
      description: Get current authenticated user information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rip_internal_app_dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rip_internal_app_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rip_internal_app_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update current user information
      parameters:
      - description: User update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rip_internal_app_dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rip_internal_app_dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rip_internal_app_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rip_internal_app_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update current user
      tags:
      - users
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"